+++ src/main/index.ts
@@ -1,5 +1,17 @@
-export * from './takers';
-export * from './code-types';
-export * from './rule';
-export * from './rule-types';
-export * from './Tokenizer';
+export {all, AllOptions} from './takers/all';
+export {char} from './takers/char';
+export {end} from './takers/end';
+export {maybe} from './takers/maybe';
+export {never} from './takers/never';
+export {none} from './takers/none';
+export {or} from './takers/or';
+export {regex} from './takers/regex';
+export {seq} from './takers/seq';
+export {skip} from './takers/skip';
+export {text, TextOptions} from './takers/text';
+export {until, UntilOptions} from './takers/until';
+export * from './takers/taker-types';
+export {Code, Var} from './code';
+export {rule} from './rules/rule';
+export {Rule, RuleHandler} from './rules/rule-types';
+export {Tokenizer} from './Tokenizer';
