+++ src/main/takers/index.ts
@@ -1,13 +1,15 @@
-export {all, AllOptions} from './all';
-export {char} from './char';
-export {end} from './end';
-export {maybe} from './maybe';
-export {never} from './never';
-export {none} from './none';
-export {or} from './or';
-export {regex} from './regex';
-export {seq} from './seq';
-export {skip} from './skip';
-export {text, TextOptions} from './text';
-export {until, UntilOptions} from './until';
+export * from './all';
+export * from './char';
+export * from './end';
+export * from './maybe';
+export * from './never';
+export * from './none';
+export * from './or';
+export * from './regex';
+export * from './seq';
+export * from './skip';
 export * from './taker-types';
+export * from './taker-utils';
+export * from './text';
+export * from './until';
+
