+++ src/test/takers/seq.test.ts
@@ -1,5 +1,5 @@
 import {InternalTaker, InternalTakerType, never, none, ResultCode, Taker, text} from '../../main';
-import {createSeqTaker, seq} from '../../main/takers/seq';
+import {createSeqTaker, seq} from '../../main/takers';
 
 describe('seq', () => {
